DataLayer-related classes were a mess. Cleaned up names and folders.
